Api completada

+++ api_rest/views.py
@@ -1,11 +1,12 @@
 from django.shortcuts import render
 from rest_framework import status
 from rest_framework.decorators import api_view, permission_classes
+from django.contrib.auth import get_user_model 
 from rest_framework.response import Response
 from rest_framework.parsers import JSONParser
 from django.views.decorators.csrf import csrf_exempt
-from .serializers import PostSerializer
-from bakaNeko.models import Post
+from .serializers import PostSerializer, ComSerializer, UserSerializer
+from bakaNeko.models import Post, Comentario, Usuario
 # Create your views here.
 
 from rest_framework.authentication import TokenAuthentication
@@ -29,6 +30,38 @@ def lista_post(request):
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+@api_view(['GET','POST'])
+@permission_classes((IsAuthenticated,))
+def lista_comentario(request):
+    if request.method == 'GET':
+        com = Comentario.objects.all()
+        serializer = ComSerializer(com, many=True)
+        return Response(serializer.data)
+    elif request.method == 'POST':
+        data = JSONParser().parse(request)
+        serializer = ComSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+@api_view(['GET','POST'])
+@permission_classes((IsAuthenticated,))
+def lista_usuario(request):
+    if request.method == 'GET':
+        user = Usuario.objects.all()
+        serializer =UserSerializer(user, many=True)
+        return Response(serializer.data)
+    elif request.method == 'POST':
+        data = JSONParser().parse(request)
+        serializer = UserSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 @api_view(['GET', 'PUT', 'DELETE'])
 @permission_classes((IsAuthenticated,))
@@ -43,7 +76,8 @@ def control_post(request, id):
         serializer = PostSerializer(p)
         return Response(serializer.data)
     if request.method == 'PUT':
-        data = JSONParser().parse(request)
+        data = JSONParser().parse(request)       
+
         serializer = PostSerializer(p, data=data)
         if serializer.is_valid():
             serializer.save()
@@ -51,5 +85,61 @@ def control_post(request, id):
         else:
             return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
     elif request.method == 'DELETE':
-        p.delete()
-        return Response(status = status.HTTP_204_NO_CONTENT)
+        if p.usuario_id == request.user.id or request.user.is_staff:
+            p.delete()
+            return Response(status = status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status = status.HTTP_401_UNAUTHORIZED)
+
+@api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes((IsAuthenticated,))
+def control_comentario(request, id):
+    try:
+        c = Comentario.objects.get(idCom = id)
+
+    except Comentario.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+    
+    if request.method == 'GET':
+        serializer = ComSerializer(c)
+        return Response(serializer.data)
+    if request.method == 'PUT':
+        data = JSONParser().parse(request)       
+
+        serializer = ComSerializer(c, data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        else:
+            return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'DELETE':
+        if c.usuario_id == request.user.id or request.user.is_staff:
+            c.delete()
+            return Response(status = status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status = status.HTTP_401_UNAUTHORIZED)
+
+@api_view(['GET', 'PUT'])
+@permission_classes((IsAuthenticated,))
+def control_usuario(request, id):
+    try:
+        u = Usuario.objects.get(id = id)
+
+    except Usuario.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+    
+    if request.method == 'GET':
+        serializer = UserSerializer(u)
+        return Response(serializer.data)
+    if request.method == 'PUT':
+        data = JSONParser().parse(request)       
+
+        serializer = UserSerializer(u, data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        else:
+            return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+    
+
+
